Where to begin... This was supposed to be an update merge to bring code written here to the same level as in the master on github; however, then all of a sudden some new functionality was added and some other code cleanup done.  Still need to merge the remaining threads into this one...
Major change is that added a orders table that is populated when an order
is placed with a SBA tracked item in the cart. Also fixed the
header_php_sba.php file.. Was missing a reset on the array generated in
the store's header_php.php file.  Still torn as to which was is the "right"
way to go with this part of the mod... By adding distinct into the original
file, that file is different than a default store, and upon upgrade may be
overwritten or may stay the same if that file is updated with the distinct
query term.  The other route, with the additional file, well, it doesn't
get updated per se and may leave behind old code that should be...
Tough to say...