Provide location of where this wonderful plugin can be found.
As a result of having to merge the filesets of this and the 1.5.4 version, a new repo was created instead of a fork off of this.  Am updating this file record to indicate where it can now be found.  The future of it is up to potteryhouse; however, there are still active references to the issues documented against this version (closed as they may be).  Been trying to address them all and or capture the remaining portions of them as applicable.  Potteryhouse with the support and help of jeking totally revived this plugin and made it possible to become a fully functional add-in.